float64

+++ lab06/functions/parsefloat.go
@@ -5,11 +5,11 @@ import (
 	"strings"
 )
 
-func ParseFloat32(s string) (float32, error) {
+func ParseFloat64(s string) (float64, error) {
 	cleanedString := strings.ReplaceAll(s, "$", "")
 	val, err := strconv.ParseFloat(cleanedString, 32)
 	if err != nil {
 		return 0, err
 	}
-	return float32(val), nil
+	return float64(val), nil
 }
